+++ packages/app/main/src/state/reducers/windowState.ts
@@ -37,11 +37,13 @@ import {
   REMEMBER_BOUNDS,
   REMEMBER_THEME,
   REMEMBER_ZOOM_LEVEL,
+  SET_AVAILABLE_THEMES,
   RememberBoundsPayload,
   RememberThemePayload,
   RememberZoomLevelPayload,
   WindowStateAction,
   WindowStatePayload,
+  SetAvailableThemesPayload,
 } from '../actions/windowStateActions';
 
 export function windowState(
@@ -69,6 +71,11 @@ export function windowState(
     case REMEMBER_THEME: {
       const { theme } = action.payload as RememberThemePayload;
       return { ...state, theme };
+    }
+
+    case SET_AVAILABLE_THEMES: {
+      const { availableThemes } = action.payload as SetAvailableThemesPayload;
+      return { ...state, availableThemes };
     }
 
     default: